requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
slack-sdk==3.26.2
openai==1.12.0
httpx==0.26.0
pydantic-settings==2.1.0
python-multipart==0.0.9
playwright==1.40.0
tweepy==4.14.0
python-dotenv==1.0.0
